+++ python/cugraph/structure/graph.py
@@ -235,7 +235,7 @@ class Graph:
                 for c in edgelist_df.columns:
                     if c in ['src', 'dst']:
                         df[c] = self.edgelist.renumber_map[edgelist_df[c]].\
-                            reset_index().drop('index')
+                            reset_index(drop=True)
                     else:
                         df[c] = edgelist_df[c]
             return df
@@ -353,9 +353,9 @@ class Graph:
         df = graph_wrapper.get_two_hop_neighbors(self)
         if self.renumbered is True:
             df['first'] = self.edgelist.renumber_map[df['first']].\
-                reset_index().drop('index')
+                reset_index(drop=True)
             df['second'] = self.edgelist.renumber_map[df['second']].\
-                reset_index().drop('index')
+                reset_index(drop=True)
         return df
 
     def number_of_vertices(self):