+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - PR #664 Added support for Louvain early termination.
 - PR #667 Drop `cython` from run requirements in conda recipe
 - PR #666 Incorporate multicolumn renumbering in python graph class for Multi(Di)Graph
+- PR #685 Avoid deep copy in index reset
 
 ## Bug Fixes
 - PR #634 renumber vertex ids passed in analytics